UGENE-1232 DNA statistics error: negative Molar ext. coef.
git-svn-id: https://local.ugene.unipro.ru/svn/ugene/trunk@2826 7ca3a678-163c-e011-b84f-001e8c0267ff

+++ src/plugins/dna_stat/src/DNAStatProfileTask.h
@@ -46,7 +46,7 @@ private:
     QVector<qint64>         contentCounter;
     //TODO: optimize
     QMap<QByteArray, int>   diNuclCounter;
-    int nA, nT, nC, nG;
+    qint64 nA, nT, nC, nG;
     QString resultText;
     // static tables
     static QMap<char,double> pMWMap; // protein molecular weight
